move resolve style input to helpers
