+++ src/components/layouts/Flex/Flex.tsx
@@ -1,5 +1,5 @@
 import { createFlexConfigs, FlexProps } from "./flex.props";
-import { resolveStyleInput } from "@/props/main.props";
+import { resolveStyleInput } from "@/helpers/resolveStyleInput";
 import { forwardRef } from "react";
 
 export const Flex = forwardRef<HTMLDivElement, FlexProps>(
