+++ src/components/layouts/Flex/flex.props.ts
@@ -16,7 +16,7 @@ import {
   rowGapVariants,
   Wrap,
 } from "./flex.css";
-import { createConfig, VariantConfig } from "@/styles/utils/resolveVariants";
+import { createConfig, VariantConfig } from "@/helpers/resolveVariants";
 import { DimensionProps } from "@/props/shared/dimension.props";
 import { PositionProps } from "@/props/shared/position.props";
 import { PaddingProps } from "@/props/shared/padding.props";
