+++ src/components/ui/Separator/Separator.tsx
@@ -1,5 +1,5 @@
 import { createSeparatorConfigs, SeparatorProps } from "./separator.props";
-import { resolveStyleInput } from "@/props/main.props";
+import { resolveStyleInput } from "@/helpers/resolveStyleInput";
 
 export const Separator: React.FC<SeparatorProps> = ({
   onClick,
