+++ src/components/ui/Separator/separator.props.ts
@@ -2,7 +2,7 @@ import { ColorProps, CustomStyleProps } from "@/components/common";
 import { MarginProps } from "@/props/shared/margin.props";
 import { OpacityProps } from "@/props/shared/opacity.props";
 import { horizontalVariants, Size, verticalVariants } from "./separator.css";
-import { createConfig, VariantConfig } from "@/styles/utils/resolveVariants";
+import { createConfig, VariantConfig } from "@/helpers/resolveVariants";
 
 type Orientation = "horizontal" | "vertical";
